delay

+++ blok2b/guitarfx/distortion.cpp
@@ -10,7 +10,7 @@ void Distortion::setGain(float gain) {
 
 
 void Distortion::process(float* buffer) {
-    for (unsigned long n=0; n<FRAMES; n++) {
+    for (int n=0; n<FRAMES; n++) {
         for (int k=0; k<CHANNELS; k++) {
             buffer[n*CHANNELS+k] = tanh(buffer[n*CHANNELS+k] * gain);
         }
